+++ server/src/main/java/com/ones/admin/common/web/ApiResourceMetadata.java
@@ -16,4 +16,8 @@ public @interface ApiResourceMetadata {
     ApiLifecycleStatus lifecycle() default ApiLifecycleStatus.UNSPECIFIED;
 
     ApiRiskLevel riskLevel() default ApiRiskLevel.UNSPECIFIED;
+
+    String sunsetVersion() default "";
+
+    String replacementApiKey() default "";
 }